+++ src/main/kotlin/com/flyn/sarcopenia_server/Main.kt
@@ -1,0 +1,7 @@
+package com.flyn.sarcopenia_server
+
+import com.flyn.sarcopenia_server.server.Server
+
+fun main() {
+    Server.start()
+}