+++ src/main/kotlin/com/flyn/sarcopenia_server/data/RequestData.kt
@@ -1,0 +1,3 @@
+package com.flyn.sarcopenia_server.data
+
+data class RequestData(var value: Int, var message: String)