+++ CHANGELOG.md
@@ -8,3 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2022-03-18
 ### Added
+- Basic server model
+- Request and response data
+- Data decoder and encoder